refactor: 스트라이크와 볼 개수 카운트 기능 함수 파일 분리
